#1016 Implemeted save action. Made security matrix master standalone. Implmented logic that manages the state of save and cancel buttons.

+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/entity/SecurityMatrixSaveAction.java
@@ -24,35 +24,35 @@ public class SecurityMatrixSaveAction  extends AbstractFunctionalEntityWithCentr
 
     @IsProperty(Object.class)
     @Title("Token Role Associations To Save")
-    private Map<String, List<Long>> associationsToSave = new HashMap<>();
+    private Map<String, List<Integer>> associationsToSave = new HashMap<>();
 
     @IsProperty(Object.class)
     @Title("Token Role Associations To Remove")
-    private Map<String, List<Long>> associationsToRemove = new HashMap<>();
+    private Map<String, List<Integer>> associationsToRemove = new HashMap<>();
 
     protected SecurityMatrixSaveAction() {
         setKey(NO_KEY);
     }
 
     @Observable
-    protected SecurityMatrixSaveAction setAssociationsToRemove(final Map<String, List<Long>> associationsToRemove) {
+    protected SecurityMatrixSaveAction setAssociationsToRemove(final Map<String, List<Integer>> associationsToRemove) {
         this.associationsToRemove.clear();
         this.associationsToRemove.putAll(associationsToRemove);
         return this;
     }
 
-    public Map<String, List<Long>> getAssociationsToRemove() {
+    public Map<String, List<Integer>> getAssociationsToRemove() {
         return Collections.unmodifiableMap(associationsToRemove);
     }
 
     @Observable
-    protected SecurityMatrixSaveAction setAssociationsToSave(final Map<String, List<Long>> aasociationsToSave) {
+    protected SecurityMatrixSaveAction setAssociationsToSave(final Map<String, List<Integer>> aasociationsToSave) {
         this.associationsToSave.clear();
         this.associationsToSave.putAll(aasociationsToSave);
         return this;
     }
 
-    public Map<String, List<Long>> getAssociationsToSave() {
+    public Map<String, List<Integer>> getAssociationsToSave() {
         return Collections.unmodifiableMap(associationsToSave);
     }
 }
